requirements.txt
fastapi
uvicorn[standard]
python-multipart
opencv-python-headless
numpy
Pillow
scipy
